Refactor BookService and repository to avoid duplicate entries

+++ src/main/java/br/com/spring/books/services/BookService.java
@@ -1,9 +1,11 @@
 package br.com.spring.books.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import br.com.spring.books.exceptions.AlreadyExistsException;
 import br.com.spring.books.exceptions.NotFoundException;
 import br.com.spring.books.models.Book;
 import br.com.spring.books.models.DTO.BookDTO;
@@ -23,6 +25,13 @@ public class BookService {
     }
 
     public Book create(BookDTO data) {
+        Optional<Book> existingBook = bookRepository
+                .findByTitleAndAuthor(data.getTitle(), data.getAuthor());
+
+        if (existingBook.isPresent()) {
+            throw new AlreadyExistsException("Este Livro já foi cadastrado.");
+        }
+
         Book book = data.toEntity();
         return this.bookRepository.save(book);
     }
@@ -48,7 +57,10 @@ public class BookService {
                 .orElseThrow(() -> new NotFoundException(id));
     }
 
-    public void delete(Long id) {
+    public void delete(Long id) throws NotFoundException {
+        if (!bookRepository.existsById(id)) {
+            throw new NotFoundException(id);
+        }
         this.bookRepository.deleteById(id);
     }
 }
